Tiny clean-up pass.

+++ src/vw/gui/TileGenerator.cc
@@ -265,8 +265,8 @@ void HttpDownloadThread::request_finished(int request_id, bool error) {
     // Now read the image out of the tempfile and save the decoded
     // pixels as the result.
     try {
-      ImageView<PixelRGBA<uint8> > vw_image = temp_tile_file.read<PixelRGBA<uint8> >();
-      //      ImageView<PixelRGBA<int16> > vw_image = temp_tile_file.read<PixelRGBA<int16> >();
+      //      ImageView<PixelRGBA<uint8> > vw_image = temp_tile_file.read<PixelRGBA<uint8> >();
+      ImageView<PixelRGBA<int16> > vw_image = temp_tile_file.read<PixelRGBA<int16> >();
       //      std::cout << "\t --> Center pixel value: " << vw_image(vw_image.cols()/2, 
       //                                                             vw_image.rows()/2) << "\n";
       buf.result = channel_cast_rescale<float>(vw_image);
